Update return types, request parameters and @RequestMapping
Return types are now objects (mostly)
ID type is now Integer instead of Long
@RequestMapping is now /object/function/singleOrBulk

+++ src/main/java/com/akshitbaunthy/restapiassignmentsolution/service/EmployeeReadService.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @Service
 public interface EmployeeReadService {
 
-    Optional<Employee> getById(Long id);
+    Optional<Employee> getById(Integer id);
     List<Employee> getAll();
 
 }
